move previous pipeline implementation to legacy2023
