+++ backend/ml/mehestan/parameters.py
@@ -1,4 +1,4 @@
-from solidago.pipeline import PipelineParameters
+from solidago.pipeline.legacy2023.parameters import PipelineParameters
 
 from tournesol.utils.constants import COMPARISON_MAX, MEHESTAN_MAX_SCALED_SCORE
