+++ solidago/src/solidago/pipeline/__init__.py
@@ -1,5 +1,3 @@
 from .inputs import TournesolInput
 from .outputs import PipelineOutput
-from .parameters import PipelineParameters
-
 from .pipeline import DefaultPipeline, Pipeline
