+++ solidago/src/solidago/pipeline/legacy2023/global_scores.py
